+++ tests/test_config_flow.py
@@ -1,14 +1,18 @@
 """Test the HaHomematic config flow."""
 from __future__ import annotations
 
+from collections.abc import Mapping
 from typing import Any
 from unittest.mock import patch
 
 from hahomematic.exceptions import AuthFailure, NoConnection
 from hahomematic.support import SystemInformation
 from homeassistant import config_entries
+from homeassistant.components import ssdp
 from homeassistant.core import HomeAssistant
-from homeassistant.data_entry_flow import RESULT_TYPE_CREATE_ENTRY, RESULT_TYPE_FORM
+from homeassistant.data_entry_flow import FlowResultType
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
 from pytest_homeassistant_custom_component.plugins import (  # noqa: F401
     enable_custom_integrations,
 )
@@ -30,14 +34,20 @@ from custom_components.homematicip_local.config_flow import (
     IF_HMIP_RF_NAME,
     IF_VIRTUAL_DEVICES_NAME,
     InvalidPassword,
+    _async_validate_config_and_get_system_information,
+    _get_instance_name,
+    _get_serial,
 )
 from custom_components.homematicip_local.const import DOMAIN
 
 from .conftest import (
+    INVALID_PASSWORD,
     TEST_ENTRY_ID,
     TEST_HOST,
     TEST_INSTANCE_NAME,
     TEST_PASSWORD,
+    TEST_SERIAL,
+    TEST_UNIQUE_ID,
     TEST_USERNAME,
 )
 
@@ -64,7 +74,7 @@ async def async_check_form(
     result = await hass.config_entries.flow.async_init(
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -73,7 +83,7 @@ async def async_check_form(
             available_interfaces=[],
             auth_enabled=False,
             https_redirect_enabled=False,
-            serial="123",
+            serial=TEST_SERIAL,
         ),
     ), patch(
         "custom_components.homematicip_local.async_setup_entry",
@@ -91,7 +101,7 @@ async def async_check_form(
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == DOMAIN
         assert result2["step_id"] == "interface"
 
@@ -107,7 +117,7 @@ async def async_check_form(
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_CREATE_ENTRY
+    assert result3["type"] == FlowResultType.CREATE_ENTRY
     assert result3["handler"] == DOMAIN
     assert result3["title"] == TEST_INSTANCE_NAME
     data = result3["data"]
@@ -120,7 +130,7 @@ async def async_check_form(
 
 async def async_check_options_form(
     hass: HomeAssistant,
-    config_entry: config_entries.ConfigEntry,
+    mock_config_entry: MockConfigEntry,
     central_data: dict[str, Any] | None = None,
     interface_data: dict[str, Any] | None = None,
 ) -> dict[str, Any]:
@@ -130,9 +140,9 @@ async def async_check_options_form(
 
     if interface_data is None:
         interface_data = {}
-    await hass.config_entries.async_add(config_entry)
-    result = await hass.config_entries.options.async_init(config_entry.entry_id)
-    assert result["type"] == RESULT_TYPE_FORM
+    mock_config_entry.add_to_hass(hass)
+    result = await hass.config_entries.options.async_init(mock_config_entry.entry_id)
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -141,7 +151,7 @@ async def async_check_options_form(
             available_interfaces=[],
             auth_enabled=False,
             https_redirect_enabled=False,
-            serial="123",
+            serial=TEST_SERIAL,
         ),
     ), patch(
         "custom_components.homematicip_local.async_setup_entry",
@@ -153,7 +163,7 @@ async def async_check_options_form(
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == TEST_ENTRY_ID
         assert result2["step_id"] == "interface"
 
@@ -169,10 +179,10 @@ async def async_check_options_form(
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_CREATE_ENTRY
+    assert result3["type"] == FlowResultType.CREATE_ENTRY
     assert result3["handler"] == TEST_ENTRY_ID
     assert result3["title"] == ""
-    return config_entry.data
+    return mock_config_entry.data
 
 
 async def test_form(hass: HomeAssistant) -> None:
@@ -193,7 +203,7 @@ async def test_options_form(
 ) -> None:
     """Test we get the form."""
     data = await async_check_options_form(
-        hass, config_entry=hmip_mock_config_entry, interface_data={}
+        hass, mock_config_entry=hmip_mock_config_entry, interface_data={}
     )
     interface = data["interface"]
     if_hmip_rf = interface[IF_HMIP_RF_NAME]
@@ -224,7 +234,7 @@ async def test_options_form_no_hmip_other_bidcos_port(
     """Test we get the form."""
     interface_data = {ATTR_HMIP_RF_ENABLED: False, ATTR_BIDCOS_RF_PORT: 5555}
     data = await async_check_options_form(
-        hass, config_entry=hmip_mock_config_entry, interface_data=interface_data
+        hass, mock_config_entry=hmip_mock_config_entry, interface_data=interface_data
     )
     interface = data["interface"]
     assert interface.get(IF_HMIP_RF_NAME) is None
@@ -249,8 +259,24 @@ async def test_form_only_hs485(hass: HomeAssistant) -> None:
     assert interface.get(IF_BIDCOS_RF_NAME) is None
     assert interface.get(IF_VIRTUAL_DEVICES_NAME) is None
 
-    if_BIDCOS_WIRED = interface[IF_BIDCOS_WIRED_NAME]
-    assert if_BIDCOS_WIRED[ATTR_PORT] == 2000
+    assert interface[IF_BIDCOS_WIRED_NAME][ATTR_PORT] == 2000
+
+
+async def test_form_only_virtual(hass: HomeAssistant) -> None:
+    """Test we get the form."""
+    interface_data = {
+        ATTR_HMIP_RF_ENABLED: False,
+        ATTR_BIDCOS_RF_ENABLED: False,
+        ATTR_VIRTUAL_DEVICES_ENABLED: True,
+        ATTR_BIDCOS_WIRED_ENABLED: False,
+    }
+    data = await async_check_form(hass, interface_data=interface_data)
+    interface = data["interface"]
+    assert interface.get(IF_HMIP_RF_NAME) is None
+    assert interface.get(IF_BIDCOS_RF_NAME) is None
+    assert interface.get(IF_BIDCOS_WIRED_NAME) is None
+
+    assert interface[IF_VIRTUAL_DEVICES_NAME][ATTR_PORT] == 9292
 
 
 async def test_form_tls(hass: HomeAssistant) -> None:
@@ -272,7 +298,7 @@ async def test_form_invalid_auth(hass: HomeAssistant) -> None:
     result = await hass.config_entries.flow.async_init(
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -293,7 +319,7 @@ async def test_form_invalid_auth(hass: HomeAssistant) -> None:
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == DOMAIN
         assert result2["step_id"] == "interface"
 
@@ -309,17 +335,17 @@ async def test_form_invalid_auth(hass: HomeAssistant) -> None:
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_FORM
+    assert result3["type"] == FlowResultType.FORM
     assert result3["errors"] == {"base": "invalid_auth"}
 
 
 async def test_options_form_invalid_auth(
-    hass: HomeAssistant, hmip_mock_config_entry: config_entries.ConfigEntry
+    hass: HomeAssistant, hmip_mock_config_entry: MockConfigEntry
 ) -> None:
     """Test we handle invalid auth."""
-    await hass.config_entries.async_add(hmip_mock_config_entry)
+    hmip_mock_config_entry.add_to_hass(hass)
     result = await hass.config_entries.options.async_init(hmip_mock_config_entry.entry_id)
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -339,7 +365,7 @@ async def test_options_form_invalid_auth(
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == TEST_ENTRY_ID
         assert result2["step_id"] == "interface"
 
@@ -355,7 +381,7 @@ async def test_options_form_invalid_auth(
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_FORM
+    assert result3["type"] == FlowResultType.FORM
     assert result3["errors"] == {"base": "invalid_auth"}
 
 
@@ -364,7 +390,7 @@ async def test_form_invalid_password(hass: HomeAssistant) -> None:
     result = await hass.config_entries.flow.async_init(
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -380,12 +406,12 @@ async def test_form_invalid_password(hass: HomeAssistant) -> None:
                 ATTR_INSTANCE_NAME: TEST_INSTANCE_NAME,
                 ATTR_HOST: TEST_HOST,
                 ATTR_USERNAME: TEST_USERNAME,
-                ATTR_PASSWORD: "In_VäLid",
+                ATTR_PASSWORD: INVALID_PASSWORD,
             },
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == DOMAIN
         assert result2["step_id"] == "interface"
 
@@ -401,17 +427,17 @@ async def test_form_invalid_password(hass: HomeAssistant) -> None:
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_FORM
+    assert result3["type"] == FlowResultType.FORM
     assert result3["errors"] == {"base": "invalid_password"}
 
 
 async def test_options_form_invalid_password(
-    hass: HomeAssistant, hmip_mock_config_entry: config_entries.ConfigEntry
+    hass: HomeAssistant, hmip_mock_config_entry: MockConfigEntry
 ) -> None:
     """Test we handle invalid auth."""
-    await hass.config_entries.async_add(hmip_mock_config_entry)
+    hmip_mock_config_entry.add_to_hass(hass)
     result = await hass.config_entries.options.async_init(hmip_mock_config_entry.entry_id)
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -426,12 +452,12 @@ async def test_options_form_invalid_password(
             {
                 ATTR_HOST: TEST_HOST,
                 ATTR_USERNAME: TEST_USERNAME,
-                ATTR_PASSWORD: "In_VäLid",
+                ATTR_PASSWORD: INVALID_PASSWORD,
             },
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == TEST_ENTRY_ID
         assert result2["step_id"] == "interface"
 
@@ -447,7 +473,7 @@ async def test_options_form_invalid_password(
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_FORM
+    assert result3["type"] == FlowResultType.FORM
     assert result3["errors"] == {"base": "invalid_password"}
 
 
@@ -456,7 +482,7 @@ async def test_form_cannot_connect(hass: HomeAssistant) -> None:
     result = await hass.config_entries.flow.async_init(
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
-    assert result["type"] == RESULT_TYPE_FORM
+    assert result["type"] == FlowResultType.FORM
     assert result["errors"] is None
 
     with patch(
@@ -477,7 +503,7 @@ async def test_form_cannot_connect(hass: HomeAssistant) -> None:
         )
         await hass.async_block_till_done()
 
-        assert result2["type"] == RESULT_TYPE_FORM
+        assert result2["type"] == FlowResultType.FORM
         assert result2["handler"] == DOMAIN
         assert result2["step_id"] == "interface"
 
@@ -493,5 +519,166 @@ async def test_form_cannot_connect(hass: HomeAssistant) -> None:
         )
         await hass.async_block_till_done()
 
-    assert result3["type"] == RESULT_TYPE_FORM
+    assert result3["type"] == FlowResultType.FORM
     assert result3["errors"] == {"base": "cannot_connect"}
+
+
+async def test_options_form_cannot_connect(
+    hass: HomeAssistant, hmip_mock_config_entry: MockConfigEntry
+) -> None:
+    """Test we handle cannot connect error."""
+    hmip_mock_config_entry.add_to_hass(hass)
+    result = await hass.config_entries.options.async_init(hmip_mock_config_entry.entry_id)
+
+    assert result["type"] == FlowResultType.FORM
+    assert result["errors"] is None
+
+    with patch(
+        "custom_components.homematicip_local.config_flow._async_validate_config_and_get_system_information",
+        side_effect=NoConnection,
+    ), patch(
+        "custom_components.homematicip_local.async_setup_entry",
+        return_value=True,
+    ):
+        result2 = await hass.config_entries.options.async_configure(
+            result["flow_id"],
+            {},
+        )
+        await hass.async_block_till_done()
+
+        assert result2["type"] == FlowResultType.FORM
+        assert result2["handler"] == TEST_ENTRY_ID
+        assert result2["step_id"] == "interface"
+
+        next(
+            flow
+            for flow in hass.config_entries.options.async_progress()
+            if flow["flow_id"] == result["flow_id"]
+        )
+
+        result3 = await hass.config_entries.options.async_configure(
+            result["flow_id"],
+            {},
+        )
+        await hass.async_block_till_done()
+
+    assert result3["type"] == FlowResultType.FORM
+    assert result3["errors"] == {"base": "cannot_connect"}
+
+
+async def test_flow_hassio_discovery(
+    hass: HomeAssistant, discovery_info: ssdp.SsdpServiceInfo
+) -> None:
+    """Test hassio discovery flow works."""
+
+    result = await hass.config_entries.flow.async_init(
+        DOMAIN,
+        data=discovery_info,
+        context={"source": config_entries.SOURCE_SSDP},
+    )
+    assert result["type"] == FlowResultType.FORM
+    assert result["step_id"] == "central"
+    assert result["description_placeholders"] is None  # {"addon": "Mock Addon"}
+
+    flows = hass.config_entries.flow.async_progress()
+    assert len(flows) == 1
+    assert flows[0].get("context", {}) == {
+        "source": "ssdp",
+        "title_placeholders": {"host": TEST_HOST, "name": TEST_INSTANCE_NAME},
+        "unique_id": TEST_UNIQUE_ID,
+    }
+
+    with patch(
+        "custom_components.homematicip_local.config_flow._async_validate_config_and_get_system_information",
+        return_value=SystemInformation(
+            available_interfaces=[],
+            auth_enabled=False,
+            https_redirect_enabled=False,
+            serial=TEST_SERIAL,
+        ),
+    ), patch(
+        "custom_components.homematicip_local.async_setup_entry",
+        return_value=True,
+    ):
+        result2 = await hass.config_entries.flow.async_configure(
+            result["flow_id"],
+            user_input={
+                ATTR_USERNAME: TEST_USERNAME,
+                ATTR_PASSWORD: TEST_PASSWORD,
+            },
+        )
+        await hass.async_block_till_done()
+        assert result2["type"] == FlowResultType.FORM
+        assert result2["handler"] == DOMAIN
+        assert result2["step_id"] == "interface"
+
+        next(
+            flow
+            for flow in hass.config_entries.flow.async_progress()
+            if flow["flow_id"] == result["flow_id"]
+        )
+
+        result3 = await hass.config_entries.flow.async_configure(
+            result["flow_id"],
+            {},
+        )
+        await hass.async_block_till_done()
+
+    assert result3["type"] == FlowResultType.CREATE_ENTRY
+    assert result3["handler"] == DOMAIN
+    assert result3["title"] == TEST_INSTANCE_NAME
+    data = result3["data"]
+    assert data[ATTR_INSTANCE_NAME] == TEST_INSTANCE_NAME
+    assert data[ATTR_HOST] == TEST_HOST
+    assert data[ATTR_USERNAME] == TEST_USERNAME
+    assert data[ATTR_PASSWORD] == TEST_PASSWORD
+
+
+async def test_hassio_discovery_existing_configuration(
+    hass: HomeAssistant,
+    hmip_mock_config_entry: MockConfigEntry,
+    discovery_info: ssdp.SsdpServiceInfo,
+) -> None:
+    """Test abort on an existing config entry."""
+    hmip_mock_config_entry.add_to_hass(hass)
+    result = await hass.config_entries.flow.async_init(
+        DOMAIN,
+        data=discovery_info,
+        context={"source": config_entries.SOURCE_SSDP},
+    )
+    assert result["type"] == FlowResultType.ABORT
+
+
+def test_config_flow_helper() -> None:
+    """Test the config flow helper."""
+
+    assert _get_instance_name(None) is None
+    assert _get_instance_name("0123456789") == "0123456789"
+    assert _get_instance_name("HomeMatic Central - test") == "test"
+    assert _get_instance_name("HomeMatic Central 0123456789") == "0123456789"
+    assert _get_serial(None) is None
+    assert _get_serial("1234") is None
+    assert _get_serial(f"9876543210{TEST_SERIAL}") == TEST_SERIAL
+
+
+async def test_async_validate_config_and_get_system_information(
+    hass: HomeAssistant, entry_data: Mapping[str, Any]
+) -> None:
+    """Test backend validation."""
+    with patch(
+        "custom_components.homematicip_local.config_flow.validate_config_and_get_system_information",
+        return_value=SystemInformation(
+            available_interfaces=[],
+            auth_enabled=False,
+            https_redirect_enabled=False,
+            serial=TEST_SERIAL,
+        ),
+    ):
+        result = await _async_validate_config_and_get_system_information(hass, entry_data)
+        assert result.serial == TEST_SERIAL
+
+    entry_data["password"] = INVALID_PASSWORD
+
+    with pytest.raises(InvalidPassword) as exc:
+        await _async_validate_config_and_get_system_information(hass, entry_data)
+    assert exc